requirements.txt
ai71==1.0.1
bm25s==0.2.10
boto3==1.35.36
datasets==3.2.0
numba==0.61.0
pandas==2.2.3
python-dotenv==1.0.1
PyStemmer==2.2.0.3
rerankers[transformers]==0.9.1.post1
sentence-transformers==3.4.1
transformers==4.47.1
transformers[torch]==4.47.1
tqdm==4.67.1
usearch==2.17.6
